In-source behave tests from OpenJDK for regression suite
The GHA job up until now relied upon running the externally-defined
behave tests for the OpenJDK container images (runtime variant).

The external test suite is primarily designed to cover the OpenJDK
image, and not to provide coverage of all of the behave-test-steps.

Rework the job to operate on a local copy of these behave tests.
(Also do a bit of spring-cleaning)

Future work can adjust the feature files to provide better coverage
for the steps defined in this repository.

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>