configurable docker API socket via DOCKER_HOST
DOCKER_HOST is an environment variable to describe the URI to connect
to.

It was originally defined and honoured by /usr/bin/docker and some bits
of the python docker API (the Client API, but not the low-level API we
use, alas). It's also supported by s2i.

BY honouring it in behave-tests-steps, the end-user can use Podman
instead of Docker, e.g.

    $ systemctl --user start podman.service
    $ export DOCKER_HOST=unix:///run/user/$(id -u)/podman/podman.sock
    $ cekit … test behave

Signed-off-by: Jonathan Dowland <jdowland@redhat.com>